requirements.txt
flask
line-bot-sdk
wit
python-dotenv
ngrok
psycopg2-binary
flask-sqlalchemy
